fix(ci): 修复 CI 编译错误并更新 workflow 配置
- 添加 `godot::` 前缀到 input_list_actions、screenshot_utils、connect_signal 等工具文件
- 修复三目运算符 String/Variant 歧义（get_info.hpp、paste_node.hpp）
- 删除未使用的变量（is_request、err 等）
- 修复括号优先级警告、sign-compare 警告
- ci.yml/release.yml: 添加 ilammy/msvc-dev-cmd@v1 为 Windows 启用 MSVC
- docs.yml: Node 升级到 22（满足 pnpm 11.5.3 的 ≥22.13 需求）

+++ extensions/src/server/registry/handler_registry.cpp
@@ -511,7 +511,7 @@ Array HandlerRegistry::search_tools(const String &query, const String &category,
 
     // Build output
     Array result;
-    for (int i = 0; i < scored.size() && i < limit; ++i) {
+    for (int i = 0; i < (int)scored.size() && i < limit; ++i) {
         Dictionary entry;
         entry["name"] = scored[i].name;
         const ToolInfo *info = find_tool_info(scored[i].name);
